Changed typescript agets to Java Script
