+++ global-news/src/app/api/news/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { agentManager } from '../../../agents/AgentManager';
+import { agentManager } from '../../../agents/AgentManager.js';
 
 export async function GET(request: NextRequest) {
   try {
